requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
Jinja2==3.1.2
requests
python-dotenv
